refactor: move out config_cluster to its own file
